09/28

+++ app/src/main/java/com/example/risingtest/src/main/store/productDetail/models/Review.kt
@@ -6,7 +6,7 @@ data class Review(
     val reviewId: Long,
     val writerIdx: Long,
     val nickname: String,
-    val score: Double,
+    val score: Float,
     val reviewPhotos: List<String>,
     val text: String,
     val createdAt: Date
